I now have the categories showing up along with their children.
1. I needed to render the categories in a seperate view (CategoryView) called from the each method in CategoriesView
2. I then had to call MenuItemsView from CategoryView.  The other ways I tried to accomplish  this were too messy and ultimtely didn't work
3. I now need to get the OrderView working again as well as hiding the list of items
