+++ scripts/views/menuItemsView.js
@@ -3,24 +3,28 @@
  */
 import {MenuItemView} from './MenuItemView'
 export default Backbone.View.extend({
+    //template: JST.categories,
     initialize: function(options){
         this.ordersCollection = options.ordersCollection;
-      this.render();
+        this.category = options.category;
+        console.log(this.category);
+        this.render();
     },
     render: function(){
-        console.log('wow',this.collection.each(groupBy('category')));
-      this.renderChildren();
+        //this.$el.append(this.template(this.category));
+        this.renderChildren();
     },
     renderChildren: function(){
         _.invoke(this.children || [], 'remove');
 
         this.children = this.collection.map(function(child) {
-            console.log(child);
+            //console.log(child);
             var view = new MenuItemView({
                 model: child,
                 ordersCollection:this.ordersCollection
             });
             this.$el.append(view.el);
+            console.log(view);
             return view;
         }.bind(this));
