+++ styles/main.scss
@@ -138,6 +138,7 @@ p{
   margin: 0;
   padding: 0;
   color: $h2-color;
+  font-size: 1.3em;
   //display: inline-block;
   //vertical-align: top;
